Move types to mapping.rs and tidy up some includes
