+++ src/lib.rs
@@ -7,15 +7,15 @@ pub mod config;
 pub mod error;
 pub mod input;
 pub mod keymap;
+pub mod mapping;
 pub mod session;
 pub mod shell;
-pub mod types;
 pub mod ui;
 
 pub use config::Config;
 pub use error::LeadrError;
+pub use mapping::{InsertType, Mapping, Mappings};
 pub use session::{LeadrSession, SessionResult};
 pub use shell::init_bash;
 pub use shell::init_zsh;
-pub use types::{Mapping, Mappings};
-pub use ui::theme::Theme;
+pub use ui::{overlay::Overlay, theme::Theme};
